# Parse protocols in Rust using the raw packets