--- a/src/module/skb/skb.rs
+++ b/src/module/skb/skb.rs
@@ -60,7 +60,6 @@ impl Collector for SkbModule {
 
         // Default list of sections.
         let mut sections: u64 = 1 << SECTION_PACKET
-            | 1 << SECTION_ETH  // FIXME.
             | 1 << SECTION_ARP
             | 1 << SECTION_IPV4
             | 1 << SECTION_IPV6