Fixes #283.

Instead of parsing protocol headers in the BPF side and reporting them through raw event sections, take advantage of the fact that we can report the (start of the) raw packet and parse it to generate the skb event (protocol) sections.

This has many advantages, summed up below:
- Using skb offsets in BPF limits what protocol sections can be reported, eg. in the IP layer in Rx we can't report TCP / UDP. Using the raw packet makes this possible as data is usually there and valid (raw packet retrieval doesn't have the same limitations regarding the skb offsets). For Tx this doesn't change much.
- This allows to have less instructions in BPF and to not report the same data twice from BPF to Rust.
- Surprisingly, performance in the Rust side only was better (parsing raw packet vs parsing multiple raw sections), see below.
- PCAP generation is now easier (raw packet is always included in the output file, no need to set specific options -some can improve the pcap generation experience but are not mandatory-).
- Easier to add support for new protocols (some will come once this is merged, that is the next step).

A few functional changes:
- Some `--skb-sections` values are now deprecated as they are always on. This is because for example we need to parse the IPv4 header anyway if we want TCP information. Copying a few fields isn't much. Only `eth` can be skipped as it's not always useful. Another reason is there are not many use cases for not reporting L3+L4 fields. If required we can add later on a limit on the max layer level to report.
- The raw packet is always reported (we can't only report the dev or ns part alone anymore). I took the opportunity to simplify things a bit since this was probably not useful.

### A note on the additional dependency

A dependency on `pnet` was added as it is supporting many protocols, is quite fast, and has ~lots of users. There was also no other crate I could find that checked all the boxes. Having said that, it's not as active as it was and the maintainer expressed their will to let someone take over (a long time ago but the issue is still opened). There was still a release in mid-2023.

Pros:
- It was easier to implement a first PR using an available crate, at least to validate the intent.
- It's not like already supported protocols in `pnet` will change a lot over the years.
- It's quite easy to replace the `pnet` use in Retis with another crate if the need arise.

Cons:
- It's not easy to find a replacement.
- Some doubts about long term efforts in `pnet`; eg. can we easily add support for parsing new protocols? Will it be maintained over time?
- `pnet` is not packaged in Fedora.

IMO things are not perfect, but are not too bad for now. It's probably acceptable. Tbf the other choice was to write a crate to parse packets myself (which I'm willing to do if the need arises or if you think we should do this from day one). If maintenance of `pnet` becomes an issue we could also vendor it, at least until we have a replacement.

### Performances of parsing events (Rust part only)

**Before**
Average first event parsing: 287µs
Average 1M events parsing: 1609080µs

**After**
Average first event parsing: 277µs
Average 1M events parsing: 1532422µs