@amorenoz 2024-03-13
Sorry I haven't got the time to look at this deeply. I would like to understand the implications. Assuming previously users likely did not enable `packet` section:
- Events files are now bigger (they contain both raw and parsed headers). This will probably take longer to write (and not super fan of adding binary data as chars in JSON :-D)
- ebpf events are the same size (1024B). But, is there more or less space? Meaning: how does packet data size compares with all the previous section events?

Am I missing something?