@atenart 2024-03-13
> Assuming previously users likely did not enable `packet` section:
> 
>     * Events files are now bigger (they contain both raw and parsed headers). This will probably take longer to write

That is right. This is not a requirement to get the functionality introduced by this PR but was more a simplification. This is handy and less error prone is various cases such as pcap generation or the upcoming python post-process.

However if there are concerns about performances writing the raw packet in the events file could still be opt-in.

> (and not super fan of adding binary data as chars in JSON :-D)

This is not adding this (pcap support did), but I get the point :) We can probably use `base64` instead (and break compatibility, not that I think it is an issue).

>     * ebpf events are the same size (1024B). But, is there more or less space? Meaning: how does packet data size compares with all the previous section events?

An Eth/IPv4/TCP packet w/o the raw packet was 57B. Now with the raw packet only it's 263B. In case the raw packet was retrieved previously, the size is now smaller. In many cases the raw packet length ending up in the events file is smaller than that. I intentionally did not try to get too smart in the BPF side about this as otherwise parsing the packet would depend on where we are in the stack.

Note that the raw packet section embedding 255B of the data is arbitrary and was chosen to be enough for encapsulation cases, but this could be revisited and we could probably chose something between 128B - 255B to save some space.