@atenart 2024-04-05
> I would like to see a performance test that actually saves and reads from file which is now considerably bigger (on the default configuration). Do you think that's easy to do with current benchmark system?

I think that's actually quite hard, the problem itself is not an easy one. We could see the impact of reading or writing more data, which can be done outside of Retis but that is not so easy I think, there are lots of things into play. Result might be similar, or a bit worst as we write more (note that with #377 raw packet is base64 encoded and is not that big).

But the real question is: is there an impact on the performance limiting part? We don't really know yet which one of the BPF<>Rust (+ parsing) part and the event printing + writing one is the limiting factor, although it's very likely the Rust side is (as it's the dequeuing/parsing side). Moreover, current benchmark commands are designed to optimize a specific part, but not really to study the impact of a change globally.

In addition, the whole skb BPF event was changed (less data) and its parsing is completely different (and moved in Rust). That also has an impact overall.

I did a test though: collection of events using the generic profile while using a high tput TCP transfer, and looked at dropped events. That's the best indicator we have IMO. Both versions (main and #377) had similar numbers (in the 300k-700k) with the main one having slightly better numbers in general.

On top of this looked at flamegraphs: the write is done by write_all -> write (blocking) and that takes only 3.5% of the total time spent.

I don't think ready/writing slightly more bits in the event file has a big impact on performances. But the overall change, with moving the packet parsing in Rust probably has. It might be more efficient in comparison, but this puts more pressure on the Rust side and IMO has an impact overall. This could be parallelized and would bring nice results, not that it's an easy topic (eg. how to order events).